Tidy character control namespacing
